backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
redis[hiredis]==5.0.7
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2
pydantic==2.8.0
pydantic-settings==2.3.4
python-multipart==0.0.9
httpx==0.27.0
minio==7.2.7
prometheus-client==0.20.0
python-dotenv==1.0.1

# Export dependencies
reportlab==4.1.0
python-docx==1.1.0
weasyprint==60.2
markdown2==2.4.13
Pillow==10.2.0

# Authentication dependencies
bcrypt==4.1.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# LangGraph and LangChain dependencies
langgraph==1.0.10
langgraph-checkpoint-postgres==3.0.4
langchain==1.2.10
langchain-openai==1.1.10
langchain-community==0.4.1
langchain-core==1.2.17